# docs/api-spec.yml
# Vid4You API contract — reset for this product (start-video-project, JOS-134,
# Decision 9). The previous version of this file described an unrelated
# inherited application (an LTI recruitment platform) end to end and has been
# fully removed.
#
# GENERATED, not hand-written: this file is produced from the Fastify + Zod
# route schemas the backend actually serves (`GET /docs/json` against a
# running instance), per `docs/backend-standards.md` § API and OpenAPI
# Conventions — "OpenAPI is generated... never hand-written and never allowed
# to drift from the actual validators." Task 5.3 confirmed this by generating
# directly from the real running server, not transcribing intent.
#
# Scope note: this story (JOS-134) only implements `POST /sessions` and
# `GET /languages`. The other paths below (`GET /sessions/{sessionId}`,
# pause/continue, scene retry/correct, downloads, the provider-callback
# simulation, the SSE `/events` stream) already exist because they were
# promoted from the `define-backend-stack` (JOS-179) skeleton this story
# extends rather than replaces — they belong to their own future stories
# (consult-session/JOS-135, US-20/US-21, US-25/US-26, US-31/US-32,
# define-live-updates/JOS-183) and are documented here only because they are
# real, not because this change built them. Regenerate this file (same
# command) whenever a route changes, rather than hand-editing it out of sync.
#
openapi: 3.0.3
info:
  title: Vid4You API
  version: 0.0.0
  description: >-
    Local, single-user API for turning a script into a narrated MP4 (Vid4You). Generated from the
    Fastify + Zod route schemas in backend/src/routes.ts (docs/backend-standards.md, API and OpenAPI
    Conventions) - never hand-written, to avoid drifting from the real validators.
components:
  schemas: {}
paths:
  /health:
    get:
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
                additionalProperties: false
  /languages:
    get:
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    code:
                      type: string
                    label:
                      type: string
                  required:
                    - code
                    - label
                  additionalProperties: false
  /sessions:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                script:
                  type: string
                language:
                  type: string
                  enum:
                    - en
                    - es
                    - fr
                    - de
                    - pt
              required:
                - title
                - script
                - language
              additionalProperties: false
      responses:
        '201':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    type: object
                    properties:
                      type:
                        type: string
                        enum:
                          - session
                      sessionId:
                        type: string
                      title:
                        type: string
                      script:
                        type: string
                      language:
                        type: string
                      state:
                        type: string
                      paused:
                        type: boolean
                      failedPhase:
                        type: string
                      updatedAt:
                        type: string
                    required:
                      - type
                      - sessionId
                      - title
                      - script
                      - language
                      - state
                      - paused
                      - updatedAt
                    additionalProperties: false
                  scenes:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum:
                            - scene
                        sessionId:
                          type: string
                        sceneId:
                          type: string
                        index:
                          type: number
                        state:
                          type: string
                        affectedStage:
                          type: string
                          enum:
                            - image
                            - video
                        errorCause:
                          type: string
                          nullable: true
                        provider:
                          type: string
                        attempts:
                          type: number
                        result:
                          type: object
                          properties:
                            imageUrl:
                              type: string
                            videoUrl:
                              type: string
                          additionalProperties: false
                        instruction:
                          type: string
                        updatedAt:
                          type: string
                      required:
                        - type
                        - sessionId
                        - sceneId
                        - index
                        - state
                        - updatedAt
                      additionalProperties: false
                required:
                  - session
                  - scenes
                additionalProperties: false
  /sessions/{sessionId}:
    get:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          in: path
          name: sessionId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  session:
                    type: object
                    properties:
                      type:
                        type: string
                        enum:
                          - session
                      sessionId:
                        type: string
                      title:
                        type: string
                      script:
                        type: string
                      language:
                        type: string
                      state:
                        type: string
                      paused:
                        type: boolean
                      failedPhase:
                        type: string
                      updatedAt:
                        type: string
                    required:
                      - type
                      - sessionId
                      - title
                      - script
                      - language
                      - state
                      - paused
                      - updatedAt
                    additionalProperties: false
                  scenes:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                          enum:
                            - scene
                        sessionId:
                          type: string
                        sceneId:
                          type: string
                        index:
                          type: number
                        state:
                          type: string
                        affectedStage:
                          type: string
                          enum:
                            - image
                            - video
                        errorCause:
                          type: string
                          nullable: true
                        provider:
                          type: string
                        attempts:
                          type: number
                        result:
                          type: object
                          properties:
                            imageUrl:
                              type: string
                            videoUrl:
                              type: string
                          additionalProperties: false
                        instruction:
                          type: string
                        updatedAt:
                          type: string
                      required:
                        - type
                        - sessionId
                        - sceneId
                        - index
                        - state
                        - updatedAt
                      additionalProperties: false
                required:
                  - session
                  - scenes
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                required:
                  - error
                additionalProperties: false
  /sessions/{sessionId}/pause:
    post:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          in: path
          name: sessionId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
  /sessions/{sessionId}/continue:
    post:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          in: path
          name: sessionId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
                additionalProperties: false
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
  /sessions/{sessionId}/scenes/{sceneId}/retry:
    post:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          in: path
          name: sessionId
          required: true
        - schema:
            type: string
            format: uuid
          in: path
          name: sceneId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
  /sessions/{sessionId}/scenes/{sceneId}/correct:
    post:
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                instruction:
                  type: string
                  minLength: 1
              required:
                - instruction
              additionalProperties: false
      parameters:
        - schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          in: path
          name: sessionId
          required: true
        - schema:
            type: string
            format: uuid
          in: path
          name: sceneId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required:
                  - ok
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
  /sessions/{sessionId}/scenes/{sceneId}/download/{kind}:
    get:
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: sessionId
          required: true
        - schema:
            type: string
            format: uuid
          in: path
          name: sceneId
          required: true
        - schema:
            type: string
            enum:
              - image
              - video
          in: path
          name: kind
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
  /sessions/{sessionId}/download/final-video:
    get:
      parameters:
        - schema:
            type: string
            pattern: ^[0-9A-HJKMNP-TV-Z]{26}$
          in: path
          name: sessionId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: string
        '404':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  reason:
                    type: string
                required:
                  - ok
                  - reason
                additionalProperties: false
  /internal/provider-callback/{requestId}:
    post:
      parameters:
        - schema:
            type: string
            format: uuid
          in: path
          name: requestId
          required: true
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  applied:
                    type: boolean
                  note:
                    type: string
                required:
                  - applied
                  - note
                additionalProperties: false
  /events:
    get:
      responses:
        '200':
          description: Default Response
  /:
    get:
      responses:
        '200':
          description: Default Response